Updated algorithm which implements fixed send time and changed controlsystem to P-controller

+++ Arduino/CalculusOperations.cpp
@@ -1,12 +1,12 @@
 #include "CalculusOperations.h"
 
-double CalculusOperations::Integrate(double newValue, double currentValue, double& deltaTime)
+double CalculusOperations::Integrate(double newValue, double currentValue, unsigned long& deltaTime)
 {	
 	currentValue += newValue * deltaTime;
 	return currentValue;
 }
 
-double CalculusOperations::Differentiate(double currentValue, double previousValue, double& deltaTime)
+double CalculusOperations::Differentiate(double currentValue, double previousValue, unsigned long& deltaTime)
 {
 	return (currentValue - previousValue) / deltaTime;
 }
